# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend API for the Kemit Trip Planner application.",
        "title": "Kemit Trip Planner API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/trips/{id}/members": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of members for a trip",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "members"
                ],
                "summary": "List members",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.TripMember"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new member to a trip",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "members"
                ],
                "summary": "Add member",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Member info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.AddMemberInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.TripMember"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/trips/{id}/members/{userId}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the role of a trip member",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "members"
                ],
                "summary": "Update member role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID (UUID)",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Role update info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.UpdateMemberRoleInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a member from a trip",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "members"
                ],
                "summary": "Remove member",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID (UUID)",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.ChatMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_verified": {
                    "type": "boolean"
                },
                "related_places": {
                    "description": "JSON-encoded string slice",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "role": {
                    "description": "\"user\" or \"assistant\"",
                    "type": "string"
                },
                "session": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ChatSession"
                        }
                    ]
                },
                "session_id": {
                    "type": "string"
                }
            }
        },
        "models.ChatSession": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ChatMessage"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.Notification": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_read": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.Place": {
            "type": "object",
            "properties": {
                "best_season": {
                    "description": "winter, summer, spring, autumn, any",
                    "type": "string"
                },
                "budget_level": {
                    "description": "low, medium, high",
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "crowd_level": {
                    "description": "quiet, moderate, crowded",
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration_needed": {
                    "description": "hours",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "reviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PlaceReview"
                    }
                },
                "saved_by": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SavedPlace"
                    }
                },
                "suitable_age": {
                    "description": "comma-separated: kid,teen,adult,senior",
                    "type": "string"
                },
                "suitable_for": {
                    "description": "comma-separated: family,couple,solo,friends",
                    "type": "string"
                },
                "thumbnail_url": {
                    "description": "externally hosted image URL",
                    "type": "string"
                },
                "trip_activities": {
                    "description": "Relations",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TripActivity"
                    }
                }
            }
        },
        "models.PlaceReview": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "place": {
                    "$ref": "#/definitions/models.Place"
                },
                "place_id": {
                    "type": "integer"
                },
                "rating": {
                    "description": "1-5",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.Preferences": {
            "type": "object",
            "properties": {
                "age_group": {
                    "description": "teen / adult / senior",
                    "type": "string"
                },
                "budget": {
                    "description": "low / medium / high",
                    "type": "string"
                },
                "crowd_preference": {
                    "description": "crowded / quiet / no_preference",
                    "type": "string"
                },
                "interests": {
                    "description": "[\"history\",\"adventure\",…]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "season": {
                    "description": "winter / spring / summer / autumn",
                    "type": "string"
                },
                "travel_companion": {
                    "description": "solo / couple / family / friends",
                    "type": "string"
                }
            }
        },
        "models.SavedPlace": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "place": {
                    "$ref": "#/definitions/models.Place"
                },
                "place_id": {
                    "type": "integer"
                },
                "user": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.Trip": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "destinations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TripDestination"
                    }
                },
                "end_date": {
                    "type": "string"
                },
                "expenses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TripExpense"
                    }
                },
                "id": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TripMember"
                    }
                },
                "preferences": {
                    "description": "same JSON shape as user preferences",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Preferences"
                        }
                    ]
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "travelers_count": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.TripActivity": {
            "type": "object",
            "properties": {
                "activity_name": {
                    "type": "string"
                },
                "activity_type": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration_hours": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "estimated_cost": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "order_in_day": {
                    "type": "integer"
                },
                "place": {
                    "$ref": "#/definitions/models.Place"
                },
                "place_id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "number"
                },
                "start_time": {
                    "type": "string"
                },
                "trip_day": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TripDay"
                        }
                    ]
                },
                "trip_day_id": {
                    "type": "string"
                }
            }
        },
        "models.TripDay": {
            "type": "object",
            "properties": {
                "activities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TripActivity"
                    }
                },
                "date": {
                    "type": "string"
                },
                "day_number": {
                    "type": "integer"
                },
                "hours_used": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "trip": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Trip"
                        }
                    ]
                },
                "trip_destination": {
                    "$ref": "#/definitions/models.TripDestination"
                },
                "trip_destination_id": {
                    "type": "string"
                },
                "trip_id": {
                    "type": "string"
                }
            }
        },
        "models.TripDestination": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "dominant category for this city",
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "days_allocated": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "order_in_trip": {
                    "type": "integer"
                },
                "trip": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Trip"
                        }
                    ]
                },
                "trip_days": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TripDay"
                    }
                },
                "trip_id": {
                    "type": "string"
                }
            }
        },
        "models.TripExpense": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "trip_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.TripMember": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "role": {
                    "description": "owner, editor, viewer",
                    "type": "string"
                },
                "trip": {
                    "$ref": "#/definitions/models.Trip"
                },
                "trip_id": {
                    "type": "string"
                },
                "user": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "auth_provider": {
                    "type": "string"
                },
                "avatar_url": {
                    "type": "string"
                },
                "badge_type": {
                    "type": "string"
                },
                "chat_sessions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ChatSession"
                    }
                },
                "city": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "explorer_points": {
                    "type": "integer"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_login": {
                    "type": "string"
                },
                "notifications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Notification"
                    }
                },
                "preferences": {
                    "$ref": "#/definitions/models.Preferences"
                },
                "reviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PlaceReview"
                    }
                },
                "role": {
                    "type": "string"
                },
                "saved_places": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SavedPlace"
                    }
                },
                "trip_members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TripMember"
                    }
                },
                "trips": {
                    "description": "Relations",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Trip"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "services.AddMemberInput": {
            "type": "object",
            "required": [
                "email",
                "role"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "editor",
                        "viewer"
                    ]
                }
            }
        },
...
        "services.UpdateMemberRoleInput": {
            "type": "object",
            "required": [
                "role"
            ],
            "properties": {
                "role": {
                    "type": "string",
                    "enum": [
                        "editor",
                        "viewer"
                    ]
                }
            }
        },
...
        "utils.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/utils.FieldError"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "utils.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "utils.Meta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/utils.ErrorBody"
                },
                "meta": {
                    "$ref": "#/definitions/utils.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}